Improved responsiveness for small screens
Adresses issue [#440 - ASCENT responsive layout](https://github.ibm.com/gsi-labs/industry-cloud-planning/issues/440).

FIXED: When the size of the browser is reduced to tablet/phone size, the left menu collapses into a hamburger menu at the top but the space that it occupied is not utilized by the application. It makes the application unusable for smaller form factors.